+++ src/main.py
@@ -8,7 +8,7 @@ import argparse
 import torch
 import math
 import faulthandler
-import BRAs_VoiceAndVision.src.tempfile as tempfile
+import uart
 from enum import Enum
 
 faulthandler.enable()
